requirements.txt
discord.py==1.7.3
python-dotenv==0.20.0
discord-components==2.1.2